pyproject.toml
[project]
name = "agentic-advisory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.5.7",
    "chromadb>=1.5.2",
    "ddgs>=9.11.2",
    "google-maps-places>=0.7.0",
    "googlemaps>=4.10.0",
    "openai>=2.24.0",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "sqlalchemy>=2.0.48",
]
